# PERF: maybe_convert_numeric speedup