@jreback 2017-04-23
the other thing is it *might* be worth it to actually try to cast once you have found an actual int (or a uint), 

```
In [13]: (ints.astype('int64', casting='unsafe') == ints).all()
Out[13]: True

In [14]: %timeit (ints.astype('int64', casting='unsafe') == ints).all()
55.5 ms ± 1.15 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

At this point it *would* raise if you encounter a float so this is ok (you can do it in a try/except).

not sure if this is totally feasible though.