@chris-b1 2017-04-24
OK, I've added some fastpath logic that will try and convert the whole array at once - have to see if anything breaks!