@TomAugspurger 2017-04-24
This should go in 0.20, correct?