@chris-b1 2017-04-24
I need to work through a couple more tests, but assuming I can get this to
work in a non-breaking way, yes I'd think this would be good for 0.20.
`to_numeric` performance with ints has regressed quite a bit since 0.19.1
(although is more functional now, to be fair)

On Mon, Apr 24, 2017 at 10:07 AM, Tom Augspurger <notifications@github.com>
wrote:

> This should go in 0.20, correct?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/pandas-dev/pandas/pull/16104#issuecomment-296698078>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AB1b_BJWgLonT_hQzRK2-WCm3Rn7XDdOks5rzLqbgaJpZM4NFkMz>
> .
>